Testing trigger for sonar
