--- backend/homework/1/core/src/main/java/org/example/Main.java
@@ -1,15 +1,0 @@
-package org.example;
-public class Main {
-    public static void main(String[] args) {
-        System.out.printf("Hello and welcome!");
-
-
-        for (int i = 1; i <= 5; i++) {
-
-            System.out.println("i = " + i);
-        }
-    }
-}
-
-
-
